[NBS-16_create_a_gravity_module] Added flat earth gravity model as well as no gravity model.
This was quite a large commit as took oppertunity to do some spring cleaning. Had to add a new function to load
uint8 to GParser which conversely made it so that had to add a string2unit8 function to GConversion.

Added an enumerator to keep track of the different models  for the gravity module. Addde a gravity_params struct
which should help keep track of different features of the module. I reckon in the future whe modelling gravity effects
it will come in handy.

Started adding the different functions for the different models. Compleing no gravity and flat earth model.

Updated the README.md to better reflect some of the changes that have gone on.