[NBS-16_create_a_gravity_module] Made changes to fix simulation of gravity module
Added a new parameter which keeps track of the number of bodies
within the simulation. This was added to GUtilities struct.

Further, there was an error where Celestial and Satellite
bodies did not have the calculate the net force functions
added. They are now added. Ran simulation and data looks reasonable.